# backend/openapi.yml
openapi: 3.0.3
info:
  title: Todo API
  description: Personal Todo application REST API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: todos
    description: Todo management
  - name: categories
    description: Category management
  - name: tags
    description: Tag management
  - name: schedules
    description: Daily planner schedule management
  - name: stats
    description: Statistics
  - name: auth
    description: Authentication

paths:
  /api/todos:
    post:
      tags: [todos]
      summary: Create a todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TodoResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

    get:
      tags: [todos]
      summary: List todos with optional filters
      parameters:
        - name: category
          in: query
          description: Filter by category ID
          required: false
          schema:
            type: integer
            format: int64
        - name: tag
          in: query
          description: Filter by tag ID
          required: false
          schema:
            type: integer
            format: int64
        - name: priority
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Priority"
        - name: completed
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          description: Keyword search on title and description
          required: false
          schema:
            type: string
        - name: dueDate
          in: query
          description: Filter by due date (ISO 8601 date) — exact match
          required: false
          schema:
            type: string
            format: date
        - name: dueDateFrom
          in: query
          description: Filter todos with dueDate >= this date (ISO 8601 date, used for calendar month range)
          required: false
          schema:
            type: string
            format: date
        - name: dueDateTo
          in: query
          description: Filter todos with dueDate <= this date (ISO 8601 date, used for calendar month range)
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/TodoResponse"

  /api/todos/{id}:
    get:
      tags: [todos]
      summary: Get a single todo
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TodoResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [todos]
      summary: Update a todo
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TodoResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [todos]
      summary: Toggle completed status of a todo
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoPatchRequest"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [todos]
      summary: Delete a todo
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"

  /api/todos/reorder:
    patch:
      tags: [todos]
      summary: Bulk reorder todos by updating positions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRequest"
      responses:
        "204":
          description: No Content

  /api/categories:
    post:
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

    get:
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/CategoryResponse"

  /api/categories/{id}:
    put:
      tags: [categories]
      summary: Update a category
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [categories]
      summary: Delete a category
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"

  /api/tags:
    post:
      tags: [tags]
      summary: Create a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TagResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

    get:
      tags: [tags]
      summary: List all tags
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/TagResponse"

  /api/tags/{id}:
    put:
      tags: [tags]
      summary: Update a tag
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TagResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [tags]
      summary: Delete a tag
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"

  /api/todo-schedules:
    get:
      tags: [schedules]
      summary: List schedules by date
      parameters:
        - name: date
          in: query
          required: true
          description: Planner date (ISO 8601 date)
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ScheduleResponse"

    post:
      tags: [schedules]
      summary: Create a schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ScheduleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /api/todo-schedules/{id}:
    put:
      tags: [schedules]
      summary: Update schedule time
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ScheduleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [schedules]
      summary: Delete a schedule
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"

  /api/auth/login:
    post:
      tags: [auth]
      summary: Login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBase"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/auth/logout:
    post:
      tags: [auth]
      summary: Logout and invalidate session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBase"

  /api/auth/me:
    get:
      tags: [auth]
      summary: Check authentication status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        type: boolean

  /api/stats:
    get:
      tags: [stats]
      summary: Get todo statistics
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponseBase"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/StatsResponse"

components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    LoginRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string

    ApiResponseBase:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true

    Priority:
      type: string
      enum: [HIGH, MEDIUM, LOW]

    CategoryCreateRequest:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
        color:
          type: string
          example: "#6366f1"

    CategoryUpdateRequest:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
        color:
          type: string
          example: "#6366f1"

    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time

    TagCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        color:
          type: string
          nullable: true
          example: "#94a3b8"

    TagUpdateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        color:
          type: string
          nullable: true
          example: "#94a3b8"

    TagResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string

    TodoPatchRequest:
      type: object
      properties:
        completed:
          type: boolean
          nullable: true

    ReorderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [id, position]
            properties:
              id:
                type: integer
                format: int64
              position:
                type: integer

    TodoRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        priority:
          $ref: "#/components/schemas/Priority"
        dueDate:
          type: string
          format: date
          nullable: true
        categoryId:
          type: integer
          format: int64
          nullable: true
        tagIds:
          type: array
          items:
            type: integer
            format: int64
          nullable: true
        completed:
          type: boolean
          nullable: true

    TodoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        priority:
          $ref: "#/components/schemas/Priority"
        dueDate:
          type: string
          format: date
          nullable: true
        category:
          allOf:
            - $ref: "#/components/schemas/CategoryResponse"
          nullable: true
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TagResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        position:
          type: integer
          description: Order within the completed/active group (0-based)

    ScheduleRequest:
      type: object
      required: [todoId, startTime, endTime, scheduleDate]
      properties:
        todoId:
          type: integer
          format: int64
        startTime:
          type: string
          format: time
          example: "10:00:00"
        endTime:
          type: string
          format: time
          example: "11:00:00"
        scheduleDate:
          type: string
          format: date
          example: "2026-06-03"

    ScheduleUpdateRequest:
      type: object
      required: [startTime, endTime]
      properties:
        startTime:
          type: string
          format: time
          example: "14:00:00"
        endTime:
          type: string
          format: time
          example: "15:30:00"

    ScheduleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        todoId:
          type: integer
          format: int64
        startTime:
          type: string
          format: time
          example: "10:00:00"
        endTime:
          type: string
          format: time
          example: "11:00:00"
        scheduleDate:
          type: string
          format: date
          example: "2026-06-03"

    StatsResponse:
      type: object
      properties:
        totalCount:
          type: integer
          format: int64
        completedCount:
          type: integer
          format: int64
        completionRate:
          type: number
          format: double
          description: Rounded to 1 decimal place (e.g. 66.7)
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryStat"
        weeklyTrend:
          type: array
          items:
            $ref: "#/components/schemas/DailyStat"
        monthlyTrend:
          type: array
          items:
            $ref: "#/components/schemas/DailyStat"

    CategoryStat:
      type: object
      properties:
        categoryName:
          type: string
        total:
          type: integer
          format: int64
        completed:
          type: integer
          format: int64

    DailyStat:
      type: object
      properties:
        date:
          type: string
          format: date
        day:
          type: string
          nullable: true
          description: Short weekday label for weekly trend (MON–SUN), null for monthly trend
          example: MON
        count:
          type: integer
          format: int64

    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              type: object
              nullable: true

  responses:
    BadRequest:
      description: Bad Request – validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized – session not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict – resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"