# backend/src/main/java/revi1337/todo/domain/schedule/service/dto/ScheduleUpdateRequest.java
package revi1337.todo.domain.schedule.service.dto;

import jakarta.validation.constraints.NotNull;

import java.time.LocalTime;

public record ScheduleUpdateRequest(
        @NotNull LocalTime startTime,
        @NotNull LocalTime endTime
) {
}